backend/requirements.txt
# requirements.txt

# Framework et serveur
fastapi==0.120.4
# uvicorn[standard]==0.38.0
uvicorn==0.38.0

# ORM et base de données
sqlalchemy==2.0.44
sqlmodel==0.0.27
pymysql==1.1.2
# mysqlclient
psycopg2-binary==2.9.11

# Validation et Pydantic
pydantic==2.12.3
email-validator==2.3.0

# Auth & password hashing
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyjwt==2.10.1
cryptography==46.0.3

# Gestion des variables d'environnement
python-dotenv==1.2.1

# MinIO et stockage d'objets
boto3==1.40.73

# Scraping et requêtes HTTP
requests==2.32.5
# playwright==1.56.0
# imageio==2.37.2

# pip install pytest pytest-cov requests python-dotenv
